requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
folium>=0.14.0
plotly>=5.15.0
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0

